Former-commit-id: 76c5a87961e5637821a4e90b5b5f7c7bad7dffc7

+++ backend/src/main/java/com/moi/anitime/api/response/animal/AnimalPreviewRes.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AnimalPreviewRes {
-    private int desertionNo;
+    private long desertionNo;
 
     private String category; // 축종
     private String detailKind;
